Add .User into all applicable requests
This updates the fetching of users and tests that it actually works.

Added exists to the template funcs to tidy up things like {{ if .User }}
which is not a valid call if the .User does exist. Now you just need to
say {{ if exists .User }} and that does what you expect.