Added conversation summary block - completes home page
Noticing some subtleties in the templates though, they really do not like
nested structs, especially when it's nested three or four deep.

I've started to smash things up but will need to flatten all of the structs
and remove nesting, and review that the structs are doing the right thing
and have only the fields that they need.

Quirk to solve: A conversation summary does vary in structure if it comes
from a search result (has 3 extra fields on the main summary and then
3 more fields to describe the last comment info).