Manual review of all structs so far
This commit is basically a review of all of the structs and a simplification of
them (readability) by reducing the amount of nesting across all structs.

Additionally removed template whitespace reduction as we don't produce an insane
amount of whitespace, it's GZIP'd, and things like CloudFlare would strip it any
way so why bother having the server TrimSpace everywhere?