Updated vendor
Additionally this required an update to the FuncMap that at this point defaults
to the one that the hugo project uses. This is unlikely to remain the case (as
there are way more funcs than we need and the project whilst licensed as Apache
isn't really designed to be usable by others independently of hugo).